added tests and update according to comments
